requirements.txt
Django==5.2.5
djangorestframework==3.16.1
reportlab==4.4.3
